+++ src/package/package.cpp
@@ -1,3 +1,21 @@
+/*
+ * newtfmt, working title, a Newton Script file reader and writer
+ * Copyright (C) 2025  Matthias Melcher
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 
 #include "package.h"
 
@@ -64,7 +82,7 @@ int Package::load(PackageBytes &p) {
   num_parts_ = p.get_uint();
   if (num_parts_ > 32)
     std::cout << "WARNING: Unlikely number of parts (" << num_parts_ << ").\n";
-  for (int i = 0; i < num_parts_; ++i) {
+  for (int i = 0; i < (int)num_parts_; ++i) {
     part_.push_back(std::make_shared<PartEntry>(i));
     part_[i]->load(p);
   }
@@ -139,7 +157,7 @@ int Package::writeAsm(std::ofstream &f) {
   f << "\t.int\t" << num_parts_ << "\t@ num_parts\n";
   f << std::endl;
   int bytes = 52;
-  for (int i = 0; i < num_parts_; ++i) {
+  for (int i = 0; i < (int)num_parts_; ++i) {
     bytes += part_[i]->writeAsm(f);
   }
   f << "@ ===== Copyright" << std::endl;
