+++ src/package/relocation_data.cpp
@@ -1,4 +1,20 @@
-
+/*
+ * newtfmt, working title, a Newton Script file reader and writer
+ * Copyright (C) 2025  Matthias Melcher
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 
 
 #include "relocation_data.h"
@@ -51,7 +67,7 @@ int RelocationData::load(PackageBytes &p) {
   page_size_ = p.get_uint();
   num_entries_ = p.get_uint();
   base_address_ = p.get_uint();
-  for (int i=0; i<num_entries_; ++i) {
+  for (int i=0; i<(int)num_entries_; ++i) {
     relocation_set_list_.push_back(RelocationSet());
     int result = relocation_set_list_[i].load(p);
     if (result != 0)
